refactor(front): Separate verification sending into component
